pyproject.toml
[project]
name = "imaginary-agents"
version = "0.1.0"
description = "A Collection of imaginary AI Agents"
authors = [
    {name = "orchidtexture"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "openai (>=1.60.1,<2.0.0)",
    "atomic-agents (>=1.0.17,<2.0.0)",
    "schedule (>=1.2.2,<2.0.0)",
    "tweepy (>=4.15.0,<5.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pyngrok (>=7.2.3,<8.0.0)",
    "pymongo (>=4.11,<5.0)",
    "pytelegrambotapi (>=4.26.0,<5.0.0)",
    "cryptography (>=44.0.0,<45.0.0)",
    "crawl4ai (>=0.5.0.post4,<0.6.0)",
    "steel-sdk (>=0.1.0b11,<0.2.0)",
    "browser-use (>=0.1.40,<0.2.0)",
    "motor (>=3.7.0,<4.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "beanie (>=1.29.0,<2.0.0)",
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "asgi-lifespan (>=2.1.0,<3.0.0)",
    "mongomock-motor (>=0.0.35,<0.0.36)",
    "fief-client[fastapi] (>=0.20.0,<0.21.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:Unused async fixture loop scope:pytest.PytestWarning"
]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%d-%m-%Y %H:%M:%S"
